+++ apps/stock-api/src/test/java/com/momentum/application/insight/MovingAverageInsightServiceTest.java
@@ -2,9 +2,9 @@ package com.momentum.application.insight;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.momentum.domain.ma.StockMovingAverage;
-import com.momentum.domain.ma.StockMovingAveragePeriod;
-import com.momentum.domain.ma.StockMovingAverageRepository;
+import com.momentum.domain.movingaverage.StockMovingAverage;
+import com.momentum.domain.movingaverage.StockMovingAveragePeriod;
+import com.momentum.domain.movingaverage.StockMovingAverageRepository;
 import com.momentum.domain.stock.Stock;
 import com.momentum.domain.stock.StockRegime;
 import com.momentum.domain.stock.StockRepository;
