+++ core/stock/src/main/java/com/momentum/domain/movingaverage/StockMovingAverageRepository.java
@@ -1,4 +1,4 @@
-package com.momentum.domain.ma;
+package com.momentum.domain.movingaverage;
 
 import com.momentum.domain.stock.Stock;
 import java.util.List;
