+++ core/stock/src/main/java/com/momentum/infrastructure/ma/StockMovingAverageJpaRepository.java
@@ -1,6 +1,6 @@
 package com.momentum.infrastructure.ma;
 
-import com.momentum.domain.ma.StockMovingAverage;
+import com.momentum.domain.movingaverage.StockMovingAverage;
 import com.momentum.domain.stock.Stock;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
